Haulers now correctly prioritize picking up dropped energy first, though I need to add a few lines of code that makes it so that only one harvester goes to a source.
Found and corrected a bug with stationary harvesters.
I was checking if their inventory was full AFTER they had just harvested.
Thus, if the container they were sitting on was full, and they harvested with a full inventory,
they would drop resources on the ground.

Working on adding a Maintainer creep to repair walls and ramparts.
Working on having towers only repair walls and rampart if they are at a critical level.